Closes #31, closes #32

+++ infrastructure/db/course_rating_schema.py
@@ -1,4 +1,4 @@
-from sqlalchemy import (Column, String, Table, MetaData, ForeignKey, Integer)
+from sqlalchemy import (Column, String, Table, MetaData, ForeignKey, Float)
 from sqlalchemy.dialects.postgresql import UUID
 import uuid
 
@@ -10,6 +10,6 @@ course_ratings = Table(
     Column('id', UUID, primary_key=True,default=uuid.uuid4),
     Column('course_id', UUID, ForeignKey('courses.id')),
     Column('user_id', UUID, ForeignKey('course_users.user_id')),
-    Column('score', Integer),
+    Column('score', Float),
     Column('opinion', String(500))
 )
